requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.4
cffi==1.14.0
colorama==0.4.3
cryptography==3.4.7
Django==3.0.5
django-cors-headers==3.4.0
django-cprofile-middleware==1.0.5
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
yapf==0.30.0
